DMEO9S12NE64: Rename files for better conformance to naming conventions
