+++ nuttx/configs/demo9s12ne64/src/m9s12_leds.c
@@ -1,7 +1,7 @@
 /****************************************************************************
- * configs/demo9s12ne64/src/up_leds.c
+ * configs/demo9s12ne64/src/m9s12_leds.c
  *
- *   Copyright (C) 2009, 2014-2015 Gregory Nutt. All rights reserved.
+ *   Copyright (C) 2009, 2013, 2015 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
  *
  * Redistribution and use in source and binary forms, with or without
@@ -39,19 +39,28 @@
 
 #include <nuttx/config.h>
 
-#include <stdint.h>
+#include <debug.h>
 
-#include <nuttx/arch.h>
 #include <nuttx/board.h>
 #include <arch/board/board.h>
 
 #include "demo9s12ne64.h"
 
-#ifdef CONFIG_ARCH_BUTTONS
-
 /****************************************************************************
  * Pre-processor Definitions
  ****************************************************************************/
+
+/* CONFIG_DEBUG_LEDS enables debug output from this file (needs CONFIG_DEBUG
+ * with CONFIG_DEBUG_VERBOSE too)
+ */
+
+#ifdef CONFIG_DEBUG_LEDS
+#  define leddbg  lldbg
+#  define ledvdbg llvdbg
+#else
+#  define leddbg(x...)
+#  define ledvdbg(x...)
+#endif
 
 /****************************************************************************
  * Private Data
@@ -66,20 +75,32 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: board_button_initialize
+ * Name: board_led_initialize
+ *
+ * Description:
+ *   Configure and initialize on-board LEDs
+ *
  ****************************************************************************/
 
-void board_button_initialize(void)
+#ifdef CONFIG_ARCH_LEDS
+void board_led_initialize(void)
 {
 }
 
 /****************************************************************************
- * Name: board_buttons
+ * Name: board_led_on
  ****************************************************************************/
 
-uint8_t board_buttons(void)
+void board_led_on(int led)
 {
-  return 0;
 }
 
-#endif /* CONFIG_ARCH_BUTTONS */
+/****************************************************************************
+ * Name: board_led_off
+ ****************************************************************************/
+
+void board_led_off(int led)
+{
+}
+
+#endif /* CONFIG_ARCH_LEDS */
